fix(daemon): drop redundant uint64 conversion in CurrentProcessFDLimit
SageOx-Session: https://sageox.ai/repo/repo_019c5812-01e9-7b7d-b5b1-321c471c9777/sessions/2026-05-03T04-18-ryan-OxCc7i/view

+++ internal/daemon/fd_limit_unix.go
@@ -17,5 +17,5 @@ func CurrentProcessFDLimit() uint64 {
 	if err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &rlim); err != nil {
 		return 0
 	}
-	return uint64(rlim.Cur)
+	return rlim.Cur
 }
